aoc-helper: move clap args to dedicated cli.rs file
